Change from Double to BigDecimal, add 'zł' to totalCost label, little clean.

+++ src/main/java/com/kodilla/restaurantfrontend/backend/BackendTableOrder.java
@@ -5,6 +5,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @AllArgsConstructor
@@ -16,7 +17,7 @@ public class BackendTableOrder {
     private String status;
     private LocalDateTime createdTime;
     private LocalDateTime closedTime;
-    private Double totalCost;
+    private BigDecimal totalCost;
     private String description;
     private String name;
 }
